backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
python-pptx==1.0.2
python-multipart==0.0.19
anthropic==0.42.0
httpx==0.28.1
aiosqlite==0.20.0
sse-starlette==2.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
defusedxml==0.7.1
pillow==11.1.0
aiofiles==24.1.0
